+++ tbot/machine/linux/machine.py
@@ -22,7 +22,7 @@ import tbot
 from tbot import machine
 from tbot.machine import channel
 from .path import Path
-from .special import Special
+from .special import Raw, Special
 from . import shell as sh
 
 Self = typing.TypeVar("Self", bound="LinuxMachine")
@@ -171,6 +171,16 @@ class LinuxMachine(machine.Machine, machine.InteractiveMachine):
         ret, _ = self.exec(*args, stdout=stdout, timeout=timeout)
         return ret == 0
 
+    def env(self, var: str) -> str:
+        """
+        Get the value of an environment variable.
+
+        :param str var: The variable's name
+        :rtype: str
+        :returns: Value of the environment variable
+        """
+        return self.exec0("printf", "%s", Raw(f'"${{{var}}}"'))
+
     def interactive(self) -> None:
         """Drop into an interactive session on this machine."""
         channel = self._obtain_channel()
@@ -218,9 +228,18 @@ class LinuxMachine(machine.Machine, machine.InteractiveMachine):
         if cmd is not None:
             channel.raw_command(cmd)
 
-    def subshell(self) -> "_SubshellContext":
+    def subshell(
+        self: Self,
+        *args: typing.Union[str, Special[Self], Path[Self]],
+        shell: typing.Optional[typing.Type[sh.Shell]] = None,
+    ) -> "_SubshellContext":
         """
         Start a subshell for isolating environment changes.
+
+        If no arguments are supplied, the shell defined for this machine
+        is used.  If arguments are given, they are expected to open
+        a shell of type ``shell`` which defaults to the shell specified
+        for this machine.
 
         **Example**::
 
@@ -232,19 +251,33 @@ class LinuxMachine(machine.Machine, machine.InteractiveMachine):
                     lh.exec0("echo", linux.Env("FOOVAR"))  # 123
 
                 lh.exec0("echo", linux.Env("FOOVAR"))  # Empty result
+
+        :param Shell shell: Shell that is started
         """
-        return _SubshellContext(self)
+        cmd = None
+        if args != []:
+            cmd = self.build_command(*args)
+        return _SubshellContext(self, shell or self.shell, cmd)
 
 
 class _SubshellContext(typing.ContextManager):
     __slots__ = ("ch", "sh")
 
-    def __init__(self, h: LinuxMachine) -> None:
+    def __init__(
+        self,
+        h: LinuxMachine,
+        shell: typing.Type[sh.Shell],
+        cmd: typing.Optional[str] = None,
+    ) -> None:
+        self.h = h
         self.ch = h._obtain_channel()
-        self.sh = h.shell
+        self.sh = shell
+        self.cmd = cmd
 
     def __enter__(self) -> None:
-        self.ch.send(f"{self.sh.name}\n")
+        cmd = self.cmd or self.sh.name
+        tbot.log_event.command(self.h.name, cmd)
+        self.ch.send(f"{cmd}\n")
         self.ch.initialize(sh=self.sh)
 
     def __exit__(self, exc_type, exc_value, traceback) -> None:  # type: ignore